refactor: extract MMR decoding to separate module and fix type cast
- Move CCITTFaxDecoder and decode_mmr_bitmap from decode_generic.rs to new decode_mmr.rs module for better code organization
- Remove unnecessary 'as u8' cast in arithmetic.rs to fix potential type issues
- Update imports in decode_generic.rs to reference the new module

This refactoring improves modularity by isolating MMR-specific decoding logic into its own file, making the codebase easier to maintain. The type cast fix ensures correct data handling in arithmetic operations.

+++ src/arithmetic.rs
@@ -88,7 +88,7 @@ impl ArithmeticDecoder {
         if self.bp >= self.data_end {
             return;
         }
-        let b = unsafe { *self.data.add(self.bp) as u8 };
+        let b = unsafe { *self.data.add(self.bp) };
         self.bp += 1;
         if b == 0xff {
             if self.bp < self.data_end {
